// src/lib/api/interfaces/updateForm.ts
export interface IEmissionForm {
  subsidiary?: string // 사업장
  year?: number // 연도
  serialNumber?: string // 일련번호
  facilityName?: string // 내부시설명
  emissionActivity?: string // 배출활동
  activityData?: string // 활동자료
  total?: number // 합계 배출량
  uncertainty?: number // 불확도
  data1: number
  data2: number
  data3: number
  data4: number
  data5: number
  data6: number
  data7: number
  data8: number
  data9: number
  data10: number
  data11: number
  data12: number
}